requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
jinja2>=3.1.4
g4f>=0.3.1.3
httpx>=0.27.0
duckduckgo-search>=6.1.0
beautifulsoup4>=4.12.3
readability-lxml>=0.8.1
trafilatura>=1.9.0
python-dotenv>=1.0.1
pydantic>=2.7.1
pydantic-settings>=2.2.1
aiofiles>=23.2.1
pyyaml>=6.0.1
openai>=1.30.0
pandas
